# docs/trading_alerts_openapi_compress.yaml
openapi: 3.0.3
info:
  title: Trading Alerts SaaS API
  version: 7.1.0
  description: |
    Commercial SaaS platform for trading alerts and signal generation with real-time chart visualization.
    **Stack:** Next.js 15 + Flask MT5 + Dual Payment (Stripe/dLocal)
    **Auth:** NextAuth.js v4 (Google OAuth + Email/Password, JWT sessions)
    **Tiers:** FREE (5 symbols × 3 timeframes = 15 combos), PRO (15 symbols × 9 timeframes = 135 combos)
    **Payment:** Stripe (international, auto-renewal, 7-day trial), dLocal (8 markets: IN, NG, PK, VN, ID, TH, ZA, TR, manual renewal)
  contact:
    name: API Support
    email: support@tradingalerts.com
  license:
    name: Proprietary
    url: https://tradingalerts.com/license
servers:
  - url: https://api.tradingalerts.com
    description: Production server
  - url: https://staging-api.tradingalerts.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/indicators/{symbol}/{timeframe}:
    get:
      tags:
        - Indicators
      summary: Get indicator data for symbol and timeframe
      description: |
        Retrieves indicator data from centralized MT5 terminal. Access is tier-based:
        FREE: 5 symbols (BTCUSD, EURUSD, USDJPY, US30, XAUUSD) × 3 timeframes (H1, H4, D1)
        PRO: 15 symbols × 9 timeframes (M5, M15, M30, H1, H2, H4, H8, H12, D1)
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Symbol'
          description: Trading symbol
        - name: timeframe
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Timeframe'
          description: Chart timeframe
        - name: bars
          in: query
          schema:
            type: integer
            minimum: 100
            maximum: 5000
            default: 1000
          description: Number of bars to retrieve
      responses:
        '200':
          description: Indicator data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorDataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Symbol not available in user's tier
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

  # ============================================
  # ALERTS ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login (includes tier information)
...
  schemas:
...
    Symbol:
      type: string
      enum:
        - AUDJPY
        - AUDUSD
        - BTCUSD
        - ETHUSD
        - EURUSD
        - GBPJPY
        - GBPUSD
        - NDX100
        - NZDUSD
        - US30
        - USDCAD
        - USDCHF
        - USDJPY
        - XAGUSD
        - XAUUSD
      description: |
        15 total symbols. FREE: BTCUSD, EURUSD, USDJPY, US30, XAUUSD. PRO: All 15.
    Timeframe:
      type: string
      enum:
        - M5
        - M15
        - M30
        - H1
        - H2
        - H4
        - H8
        - H12
        - D1
      description: |
        9 total timeframes. FREE: H1, H4, D1. PRO: All 9 (M5 and H12 are PRO-only).
...
    IndicatorDataResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            ohlc:
              type: array
              items:
                type: object
                properties:
                  time:
                    type: integer
                  open:
                    type: number
                  high:
                    type: number
                  low:
                    type: number
                  close:
                    type: number
            horizontal:
              type: object
              properties:
                peak_1:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                peak_2:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                peak_3:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                bottom_1:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                bottom_2:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                bottom_3:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
            diagonal:
              type: object
              properties:
                ascending_1:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                ascending_2:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                ascending_3:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                descending_1:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                descending_2:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
                descending_3:
                  type: array
                  items:
                    $ref: '#/components/schemas/LinePoint'
            fractals:
              type: object
              properties:
                peaks:
                  type: array
                  items:
                    $ref: '#/components/schemas/FractalPoint'
                bottoms:
                  type: array
                  items:
                    $ref: '#/components/schemas/FractalPoint'
        cached:
          type: boolean
    LinePoint:
      type: object
      properties:
        time:
          type: integer
        value:
          type: number
    FractalPoint:
      type: object
      properties:
        time:
          type: integer
        price:
          type: number
        type:
          type: string
          enum: [peak, bottom]

    # ============================================
    # ALERT SCHEMAS
    # ============================================
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true
        details:
          type: object
          nullable: true
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...